Move service types from Commands/ to Inspectors/
- Move ApiServices, LibraryMetadataService, TypeSearchService, SignatureParser
- Update namespace from DotnetInspector.Commands to DotnetInspector.Inspectors
- Add using DotnetInspector.Inspectors to SamplesCommand and MemberTableFormatter
- These types perform inspection/extraction work, not command orchestration
