+++ src/dotnet-inspect/Commands/SamplesCommand.cs
@@ -1,4 +1,5 @@
 using System.Text;
+using DotnetInspector.Inspectors;
 using DotnetInspector.Metadata;
 using DotnetInspector.Options;
 using DotnetInspector.Output;
